Created main.py and intro sequence.
Created main function and intro. Intro asks if you have an acount, if not it allows you to create an acount file, given that the file name is unique, and if you have an acount, it allows you to sign in by name aa well as greeting you by name using line 1 of the file.